Moved propensity and relative abundance to the animal file
