+++ src/biosim/cell.py
@@ -194,8 +194,8 @@ class Cell:
         for type_of_animal, animal_list in self.animal_classes.items():
             remove_list = set()
             for index, animal in enumerate(animal_list):
-                if animal.determine_to_move():
-                    move_prob = self.compute_move_prob(animal, neighbour_cells)
+                if animal.determine_to_move() and not animal.has_migrated:
+                    move_prob = animal.compute_move_prob(neighbour_cells)
                     if sum(move_prob) == 0:
                         break
                     else:
@@ -203,6 +203,7 @@ class Cell:
                             neighbour_cells, weights=move_prob
                         )[0]
                         chosen_cell.insert_animal([animal])
+                        animal.has_migrated = True
                         remove_list.add(index)
             self.remove_multiple_animals(type_of_animal, remove_list)
 
@@ -218,132 +219,133 @@ class Cell:
         for type_of_animal, animal_list in self.animal_classes.items():
             remove_list = set()
             for index, animal in enumerate(animal_list):
+                animal.has_migrated = False  # Prepares for the next cycle
                 if animal.determine_death():
                     remove_list.add(index)
             self.remove_multiple_animals(type_of_animal, remove_list)
 
-    @staticmethod
-    def compute_move_prob(animal_type, neighbour_cells):
-        """
+    # @staticmethod
+    # def compute_move_prob(animal_type, neighbour_cells):
+    #     """
+    #
+    #     Computes the probability to move to each neighbouring cell
+    #
+    #     Parameters
+    #     ----------
+    #     animal_type: class instance
+    #                  Takes in a specific class instance
+    #     neighbour_cells:
+    #                 Takes in the neighbouring cells of the cell the animal
+    #                 class instance is located at
+    #
+    #     Returns
+    #     -------
+    #
+    #     list
+    #         List containing probability to move to each cell
+    #
+    #
+    #     """
+    #     cell_propensity = []
+    #     for cell in neighbour_cells:
+    #         propensity_cell = cell.propensity(animal_type)
+    #         cell_propensity.append(propensity_cell)
+    #
+    #     total_propensity = sum(cell_propensity)
+    #
+    #     computed_propensities = []
+    #     for cell_prop in cell_propensity:
+    #         try:
+    #             prob = cell_prop / total_propensity
+    #         except ZeroDivisionError:
+    #             prob = 0
+    #         computed_propensities.append(prob)
+    #     return computed_propensities
 
-        Computes the probability to move to each neighbouring cell
+    # def compute_relative_abundance(self, animal_class):
+    #     r"""
+    #
+    #     Computes the relative abundance for either herbivore or carnivore,
+    #     depending on the specie type.
+    #
+    #     The relative abundance is computed through this formula:
+    #
+    #     .. math::
+    #         \epsilon_{k} = \frac{f_{k}}{(n_{k}+1)F^{\text{'}}}
+    #
+    #     Where :math:`\epsilon_{k}` is the relative abundance. :math:`f_{k}` is
+    #     the current fodder for cell k, which is different for carnivores
+    #     and herbivores. :math:`n_{k}` is the amount of same species in cell
+    #     k, and :math:`F^{\text{'}}` is how much food the animal wants to eat.
+    #
+    #     Parameters
+    #     ----------
+    #     animal_class: class instance
+    #                 The class instance one needs to calculate relative
+    #                 abundance for
+    #
+    #     Returns
+    #     -------
+    #     float
+    #         The relative abundance of the current cell
+    #
+    #             """
+    #
+    #     animal_name = type(animal_class).__name__
+    #     amount_same_spec = len(self.animal_classes[animal_name])
+    #     food_wanting = animal_class.param["F"]
+    #     curr_fod = 0
+    #     if animal_name == "Herbivore":
+    #         curr_fod = self.current_fodder
+    #     elif animal_name == "Carnivore":
+    #         herb_weight_list = [
+    #             herbivore.weight
+    #             for herbivore in self.animal_classes["Herbivore"]
+    #         ]
+    #         curr_fod = sum(herb_weight_list)
+    #
+    #     if curr_fod == 0:
+    #         return 0
+    #     else:
+    #         return curr_fod / ((amount_same_spec + 1) * food_wanting)
 
-        Parameters
-        ----------
-        animal_type: class instance
-                     Takes in a specific class instance
-        neighbour_cells:
-                    Takes in the neighbouring cells of the cell the animal
-                    class instance is located at
-
-        Returns
-        -------
-
-        list
-            List containing probability to move to each cell
-
-
-        """
-        total_propensity = 0
-        cell_propensity = []
-        for cell in neighbour_cells:
-            propenisty_cell = cell.propensity(animal_type)
-            cell_propensity.append(propenisty_cell)
-            total_propensity += propenisty_cell
-
-        computed_propensities = []
-        for cell_prop in cell_propensity:
-            try:
-                prob = cell_prop / total_propensity
-            except ZeroDivisionError:
-                prob = 0
-            computed_propensities.append(prob)
-        return computed_propensities
-
-    def compute_relative_abundance(self, animal_class):
-        r"""
-
-        Computes the relative abundance for either herbivore or carnivore,
-        depending on the specie type.
-
-        The relative abundance is computed through this formula:
-
-        .. math::
-            \epsilon_{k} = \frac{f_{k}}{(n_{k}+1)F^{\text{'}}}
-
-        Where :math:`\epsilon_{k}` is the relative abundance. :math:`f_{k}` is
-        the current fodder for cell k, which is different for carnivores
-        and herbivores. :math:`n_{k}` is the amount of same species in cell
-        k, and :math:`F^{\text{'}}` is how much food the animal wants to eat.
-
-        Parameters
-        ----------
-        animal_class: class instance
-                    The class instance one needs to calculate relative
-                    abundance for
-
-        Returns
-        -------
-        float
-            The relative abundance of the current cell
-
-                """
-
-        animal_name = type(animal_class).__name__
-        amount_same_spec = len(self.animal_classes[animal_name])
-        food_wanting = animal_class.param["F"]
-        curr_fod = 0
-        if animal_name == "Herbivore":
-            curr_fod = self.current_fodder
-        elif animal_name == "Carnivore":
-            herb_weight_list = [
-                herbivore.weight
-                for herbivore in self.animal_classes["Herbivore"]
-            ]
-            curr_fod = sum(herb_weight_list)
-
-        if curr_fod == 0:
-            return 0
-        else:
-            return curr_fod / ((amount_same_spec + 1) * food_wanting)
-
-    def propensity(self, specie):
-        r"""
-
-        Computes and returns the propensity to move, the relative abundance is
-        calculated through the 'compute_relative_abundance' function.
-        The formula for propensity is given by:
-
-        .. math::
-            \pi_{i\rightarrow j} =
-            \begin{cases}
-            0 & \text{if } j \text{is Mountain or Ocean}\\
-            e^{\lambda \epsilon_{j}} & \text{otherwise}
-            \end{cases}
-
-
-        Parameters
-        ----------
-        specie: class instance
-            The class instance of animal to be used in the function
-
-        Returns
-        -------
-        float
-            The propensity to move
-
-        """
-        name = type(specie).__name__
-        cell_name = type(self).__name__
-        if (name == "Herbivore" or name == "Carnivore") and (
-            cell_name == "Ocean" or cell_name == "Mountain"
-        ):
-            return 0
-
-        relative_abundance = self.compute_relative_abundance(specie)
-        lambda_specie = specie.param["lambda"]
-
-        return math.exp(lambda_specie * relative_abundance)
+    # def propensity(self, specie):
+    #     r"""
+    #
+    #     Computes and returns the propensity to move, the relative abundance is
+    #     calculated through the 'compute_relative_abundance' function.
+    #     The formula for propensity is given by:
+    #
+    #     .. math::
+    #         \pi_{i\rightarrow j} =
+    #         \begin{cases}
+    #         0 & \text{if } j \text{is Mountain or Ocean}\\
+    #         e^{\lambda \epsilon_{j}} & \text{otherwise}
+    #         \end{cases}
+    #
+    #
+    #     Parameters
+    #     ----------
+    #     specie: class instance
+    #         The class instance of animal to be used in the function
+    #
+    #     Returns
+    #     -------
+    #     float
+    #         The propensity to move
+    #
+    #     """
+    #     name = type(specie).__name__
+    #     cell_name = type(self).__name__
+    #     if (name == "Herbivore" or name == "Carnivore") and (
+    #         cell_name == "Ocean" or cell_name == "Mountain"
+    #     ):
+    #         return 0
+    #
+    #     relative_abundance = self.compute_relative_abundance(specie)
+    #     lambda_specie = specie.param["lambda"]
+    #
+    #     return math.exp(lambda_specie * relative_abundance)
 
     def aging(self):
         """
@@ -389,7 +391,7 @@ class Cell:
 
         """
         for animal_class in animal_list:
-            if type(self).__name__ not in Animal.allowed_ladndscape:
+            if type(self).__name__ not in Animal.allowed_landscape:
                 raise TypeError(f"This cell is inhabitable.")
             self.animal_classes[type(animal_class).__name__].append(
                 animal_class
@@ -416,7 +418,7 @@ class Cell:
             if animal_name not in self.allowed_species.keys():
                 raise ValueError(f"The animal type is not allowed")
 
-            if cell_name not in Animal.allowed_ladndscape:
+            if cell_name not in Animal.allowed_landscape:
                 raise ValueError(
                     f"This cell is inhabitable for specie: {animal_name}"
                 )
